Avoid heavy auth marker import for request SecretRef markers
